requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.2.1
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
Markdown==3.5.2
model-bakery==1.17.0
packaging==23.2
prompt-toolkit==3.0.43
psycopg==3.1.18
psycopg-binary==3.1.18
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
requests==2.32.3
rpds-py==0.17.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
pyotp==2.9.0
pillow==10.1.0
faker==24.2.0 #
django-ratelimit==4.1.0
python-magic==0.4.27
firebase-admin==6.5.0
pycountry==23.12.11
sortedcontainers==2.4.0
ulid-py==1.1.0
boto3==1.34.144
django-storages==1.14.4
cryptography==43.0.0
bleach==6.1.0
mongoengine==0.28.2
pymongo==4.8.0
phonenumbers==8.13.42
openai==1.76.0
pydantic==2.9.2
ipykernel==6.29.5
django-jazzmin==3.0.1
django-import-export==4.3.3
channels==4.2.2
channels_redis==4.2.1
deepseek_ai==0.0.1
daphne>=4.0.0
python-docx==1.1.2
gevent==25.4.2
jdatetime==5.2.0
PyMuPDF==1.26.6
mutagen==1.47.0
